+++ src/apis/country/country.ts
@@ -1,0 +1,9 @@
+import { IContinent } from '../continent';
+
+export interface ICountry {
+  id: number;
+  nameKor: string;
+  nameEng: string;
+  imageUrl: string;
+  continent?: IContinent;
+}
