+++ src/apis/country/index.ts
@@ -1,0 +1,2 @@
+export type { ICountry } from './country';
+export { getCountries } from './get-countries';
